restructure interfaces
