+++ src/ui/engine/EngineRenderer.ts
@@ -6,7 +6,8 @@ import {default as ABodyRenderer} from '../ABodyRenderer';
 import DataProvider, {default as ADataProvider} from '../../provider/ADataProvider';
 import {default as Column, ICategoricalStatistics, IFlatColumn, IStatistics} from '../../model/Column';
 import {createDOM, createDOMGroup} from '../../renderer';
-import {default as RenderColumn, IGroupData, IGroupItem, IRankingBodyContext, isGroup} from './RenderColumn';
+import {default as RenderColumn} from './RenderColumn';
+import {IGroupData, IGroupItem, IRankingBodyContext, isGroup} from './interfaces';
 import EngineRankingRenderer from './EngineRankingRenderer';
 import {ILineUpRenderer} from '../index';
 import {ILineUpConfig, IRenderingOptions} from '../../lineup';
