+++ src/ui/engine/RenderColumn.ts
@@ -2,64 +2,11 @@
  * Created by Samuel Gratzl on 25.07.2017.
  */
 import {IColumn} from 'lineupengine/src';
-import Column, {ICategoricalStatistics, IStatistics} from '../../model/Column';
-import {IDataProvider, IDataRow} from '../../provider/ADataProvider';
-import {IFilterDialog} from '../../dialogs/AFilterDialog';
-import {createSummary, createToolbar, dragWidth, handleDnD} from './header';
-import {INumberColumn} from '../../model/NumberColumn';
-import {ICategoricalColumn} from '../../model/CategoricalColumn';
+import Column from '../../model/Column';
+import {createToolbar, dragWidth, handleDnD, updateHeader} from './header';
 import {IDOMCellRenderer, IDOMGroupRenderer} from '../../renderer/IDOMCellRenderers';
-import {IDOMRenderContext} from '../../renderer/RendererContexts';
-import {IGroup} from '../../model/Group';
+import {IRankingContext} from './interfaces';
 
-export interface IRankingHeaderContextContainer {
-  readonly idPrefix: string;
-  provider: IDataProvider;
-  linkTemplates: string[];
-
-  searchAble(col: Column): boolean;
-
-  autoRotateLabels: boolean;
-  filters: { [type: string]: IFilterDialog };
-
-  statsOf(col: (INumberColumn | ICategoricalColumn) & Column): ICategoricalStatistics | IStatistics | null;
-}
-
-export interface IGroupItem extends IDataRow {
-  group: IGroup;
-  relativeIndex: number;
-}
-
-export interface IGroupData extends IGroup {
-  rows: IDataRow[];
-}
-
-export function isGroup(item: IGroupData|IGroupItem): item is IGroupData {
-  return (<IGroupData>item).name !== undefined; // use .name as separator
-}
-
-export interface IRankingBodyContext extends IRankingHeaderContextContainer, IDOMRenderContext {
-  isGroup(index: number): boolean;
-  getGroup(index: number): IGroupData;
-  getRow(index: number): IGroupItem;
-}
-
-export declare type IRankingHeaderContext = Readonly<IRankingHeaderContextContainer>;
-
-export declare type IRankingContext = Readonly<IRankingBodyContext>;
-
-
-/**
- * utility function to generate the tooltip text with description
- * @param col the column
- */
-export function toFullTooltip(col: { label: string, description?: string }) {
-  let base = col.label;
-  if (col.description != null && col.description !== '') {
-    base += `\n${col.description}`;
-  }
-  return base;
-}
 
 export interface IRenderers {
   singleId: string;
@@ -105,16 +52,8 @@ export default class RenderColumn implements IColumn {
 
 
   updateHeader(node: HTMLElement, ctx: IRankingContext) {
-    node.querySelector('div.lu-label')!.innerHTML = this.c.label;
-    node.title = toFullTooltip(this.c);
     node.className = `${this.c.cssClass ? ` ${this.c.cssClass}` : ''}${(this.c.getCompressed() ? ' lu-compressed' : '')} ${this.c.headerCssClass}${ctx.autoRotateLabels ? ' rotateable' : ''}${this.c.isFiltered() ? ' lu-filtered' : ''}`;
-    const sort = <HTMLElement>node.querySelector('.lu-sort')!;
-    const {asc, priority} = this.c.isSortedByMe();
-    sort.dataset.sort = asc;
-    sort.dataset.priority = priority;
-
-
-    createSummary(<HTMLElement>node.querySelector('div.lu-summary')!, this.c, ctx);
+    updateHeader(node, this.c, ctx);
   }
 
   createCell(index: number, document: Document, ctx: IRankingContext) {
